+++ mkl_fft/interfaces/__init__.py
@@ -23,4 +23,11 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-from . import numpy_fft, scipy_fft
+from . import numpy_fft
+
+try:
+    import scipy.fft
+except ImportError:
+    pass
+else:
+    from . import scipy_fft