+++ mkl_fft/interfaces/_scipy_fft.py
@@ -32,7 +32,7 @@ in the backend.
 import contextlib
 import contextvars
 import operator
-import os
+from numbers import Number
 
 import mkl
 import numpy as np
@@ -66,31 +66,13 @@ __all__ = [
 ]
 
 
-class _cpu_max_threads_count:
-    def __init__(self):
-        self.cpu_count = None
-        self.max_threads_count = None
-
-    def get_cpu_count(self):
-        if self.cpu_count is None:
-            max_threads = self.get_max_threads_count()
-            self.cpu_count = max_threads
-        return self.cpu_count
-
-    def get_max_threads_count(self):
-        if self.max_threads_count is None:
-            # pylint: disable=no-member
-            self.max_threads_count = mkl.get_max_threads()
-
-        return self.max_threads_count
-
-
 class _workers_data:
     def __init__(self, workers=None):
-        if workers:
-            self.workers_ = workers
+        if workers is not None:  # workers = 0 should be handled
+            self.workers_ = _workers_to_num_threads(workers)
         else:
-            self.workers_ = _cpu_max_threads_count().get_cpu_count()
+            # Unlike SciPy, the default value is maximum number of threads
+            self.workers_ = mkl.get_max_threads()  # pylint: disable=no-member
         self.workers_ = operator.index(self.workers_)
 
     @property
@@ -108,8 +90,9 @@ _workers_global_settings = contextvars.ContextVar(
 
 
 def _workers_to_num_threads(w):
-    """Handle conversion of workers to a positive number of threads in the
-    same way as scipy.fft.helpers._workers.
+    """
+    Handle conversion of workers to a positive number of threads in the
+    same way as scipy.fft._pocketfft.helpers._workers.
     """
     if w is None:
         return _workers_global_settings.get().workers
@@ -117,12 +100,13 @@ def _workers_to_num_threads(w):
     if _w == 0:
         raise ValueError("Number of workers must not be zero")
     if _w < 0:
-        ub = os.cpu_count()
-        _w += ub + 1
+        # SciPy uses os.cpu_count()
+        _cpu_count = mkl.get_max_threads()  # pylint: disable=no-member
+        _w += _cpu_count + 1
         if _w <= 0:
             raise ValueError(
-                "workers value out of range; got {}, must not be"
-                " less than {}".format(w, -ub)
+                f"workers value out of range; got {w}, must not be less "
+                f"than {-_cpu_count}"
             )
     return _w
 
@@ -134,14 +118,16 @@ class _Workers:
 
     def __enter__(self):
         try:
+            # mkl.set_num_threads_local sets the number of threads to the
+            # given input number, and returns the previous number of threads
             # pylint: disable=no-member
             self.prev_num_threads = mkl.set_num_threads_local(self.n_threads)
         except Exception as e:
             raise ValueError(
-                "Class argument {} result in invalid number of threads {}".format(
-                    self.workers, self.n_threads
-                )
+                f"Class argument {self.workers} results in invalid number of "
+                f"threads {self.n_threads}"
             ) from e
+        return self
 
     def __exit__(self, *args):
         # restore old value
@@ -156,30 +142,65 @@ def _check_plan(plan):
         )
 
 
-def _check_overwrite_x(overwrite_x):
-    if overwrite_x:
-        raise NotImplementedError(
-            "Overwriting the content of `x` is currently not supported"
-        )
+# copied from scipy.fft._pocketfft.helper
+# https://github.com/scipy/scipy/blob/main/scipy/fft/_pocketfft/helper.py
+def _iterable_of_int(x, name=None):
+    if isinstance(x, Number):
+        x = (x,)
+
+    try:
+        x = [operator.index(a) for a in x]
+    except TypeError as e:
+        name = name or "value"
+        raise ValueError(
+            f"{name} must be a scalar or iterable of integers"
+        ) from e
+
+    return x
 
 
-def _cook_nd_args(x, s=None, axes=None, invreal=False):
-    if s is None:
-        shapeless = True
-        if axes is None:
-            s = list(x.shape)
-        else:
-            s = np.take(x.shape, axes)
+# copied and modified from scipy.fft._pocketfft.helper
+# https://github.com/scipy/scipy/blob/main/scipy/fft/_pocketfft/helper.py
+def _init_nd_shape_and_axes(x, shape, axes, invreal=False):
+    noshape = shape is None
+    noaxes = axes is None
+
+    if not noaxes:
+        axes = _iterable_of_int(axes, "axes")
+        axes = [a + x.ndim if a < 0 else a for a in axes]
+
+        if any(a >= x.ndim or a < 0 for a in axes):
+            raise ValueError("axes exceeds dimensionality of input")
+        if len(set(axes)) != len(axes):
+            raise ValueError("all axes must be unique")
+
+    if not noshape:
+        shape = _iterable_of_int(shape, "shape")
+
+        if axes and len(axes) != len(shape):
+            raise ValueError(
+                "when given, axes and shape arguments"
+                " have to be of the same length"
+            )
+        if noaxes:
+            if len(shape) > x.ndim:
+                raise ValueError("shape requires more axes than are present")
+            axes = range(x.ndim - len(shape), x.ndim)
+
+        shape = [x.shape[a] if s == -1 else s for s, a in zip(shape, axes)]
+    elif noaxes:
+        shape = list(x.shape)
+        axes = range(x.ndim)
     else:
-        shapeless = False
-    s = list(s)
-    if axes is None:
-        axes = list(range(-len(s), 0))
-    if len(s) != len(axes):
-        raise ValueError("Shape and axes have different lengths.")
-    if invreal and shapeless:
-        s[-1] = (x.shape[axes[-1]] - 1) * 2
-    return s, axes
+        shape = [x.shape[a] for a in axes]
+
+    if noshape and invreal:
+        shape[-1] = (x.shape[axes[-1]] - 1) * 2
+
+    if any(s < 1 for s in shape):
+        raise ValueError(f"invalid number of data points ({shape}) specified")
+
+    return tuple(shape), list(axes)
 
 
 def _validate_input(x):
@@ -301,7 +322,7 @@ def fftn(
     """
     _check_plan(plan)
     x = _validate_input(x)
-    s, axes = _cook_nd_args(x, s, axes)
+    s, axes = _init_nd_shape_and_axes(x, s, axes)
     fsc = _compute_fwd_scale(norm, s, x.shape)
 
     with _Workers(workers):
@@ -328,7 +349,7 @@ def ifftn(
     """
     _check_plan(plan)
     x = _validate_input(x)
-    s, axes = _cook_nd_args(x, s, axes)
+    s, axes = _init_nd_shape_and_axes(x, s, axes)
     fsc = _compute_fwd_scale(norm, s, x.shape)
 
     with _Workers(workers):
@@ -345,17 +366,13 @@ def rfft(
 
     For full documentation refer to `scipy.fft.rfft`.
 
-    Limitation
-    -----------
-    The kwarg `overwrite_x` is only supported with its default value.
-
     """
     _check_plan(plan)
-    _check_overwrite_x(overwrite_x)
     x = _validate_input(x)
     fsc = _compute_fwd_scale(norm, n, x.shape[axis])
 
     with _Workers(workers):
+        # Note: overwrite_x is not utilized
         return mkl_fft.rfft(x, n=n, axis=axis, fwd_scale=fsc)
 
 
@@ -367,17 +384,13 @@ def irfft(
 
     For full documentation refer to `scipy.fft.irfft`.
 
-    Limitation
-    -----------
-    The kwarg `overwrite_x` is only supported with its default value.
-
     """
     _check_plan(plan)
-    _check_overwrite_x(overwrite_x)
     x = _validate_input(x)
     fsc = _compute_fwd_scale(norm, n, 2 * (x.shape[axis] - 1))
 
     with _Workers(workers):
+        # Note: overwrite_x is not utilized
         return mkl_fft.irfft(x, n=n, axis=axis, fwd_scale=fsc)
 
 
@@ -395,10 +408,6 @@ def rfft2(
     Compute the 2-D discrete Fourier Transform for real input.
 
     For full documentation refer to `scipy.fft.rfft2`.
-
-    Limitation
-    -----------
-    The kwarg `overwrite_x` is only supported with its default value.
 
     """
     return rfftn(
@@ -427,10 +436,6 @@ def irfft2(
 
     For full documentation refer to `scipy.fft.irfft2`.
 
-    Limitation
-    -----------
-    The kwarg `overwrite_x` is only supported with its default value.
-
     """
     return irfftn(
         x,
@@ -458,18 +463,14 @@ def rfftn(
 
     For full documentation refer to `scipy.fft.rfftn`.
 
-    Limitation
-    -----------
-    The kwarg `overwrite_x` is only supported with its default value.
-
     """
     _check_plan(plan)
-    _check_overwrite_x(overwrite_x)
     x = _validate_input(x)
-    s, axes = _cook_nd_args(x, s, axes)
+    s, axes = _init_nd_shape_and_axes(x, s, axes)
     fsc = _compute_fwd_scale(norm, s, x.shape)
 
     with _Workers(workers):
+        # Note: overwrite_x is not utilized
         return mkl_fft.rfftn(x, s, axes, fwd_scale=fsc)
 
 
@@ -488,18 +489,14 @@ def irfftn(
 
     For full documentation refer to `scipy.fft.irfftn`.
 
-    Limitation
-    -----------
-    The kwarg `overwrite_x` is only supported with its default value.
-
     """
     _check_plan(plan)
-    _check_overwrite_x(overwrite_x)
     x = _validate_input(x)
-    s, axes = _cook_nd_args(x, s, axes, invreal=True)
+    s, axes = _init_nd_shape_and_axes(x, s, axes, invreal=True)
     fsc = _compute_fwd_scale(norm, s, x.shape)
 
     with _Workers(workers):
+        # Note: overwrite_x is not utilized
         return mkl_fft.irfftn(x, s, axes, fwd_scale=fsc)
 
 
@@ -512,13 +509,8 @@ def hfft(
 
     For full documentation refer to `scipy.fft.hfft`.
 
-    Limitation
-    -----------
-    The kwarg `overwrite_x` is only supported with its default value.
-
     """
     _check_plan(plan)
-    _check_overwrite_x(overwrite_x)
     x = _validate_input(x)
     norm = _swap_direction(norm)
     x = np.array(x, copy=True)
@@ -526,6 +518,7 @@ def hfft(
     fsc = _compute_fwd_scale(norm, n, 2 * (x.shape[axis] - 1))
 
     with _Workers(workers):
+        # Note: overwrite_x is not utilized
         return mkl_fft.irfft(x, n=n, axis=axis, fwd_scale=fsc)
 
 
@@ -537,18 +530,14 @@ def ihfft(
 
     For full documentation refer to `scipy.fft.ihfft`.
 
-    Limitation
-    -----------
-    The kwarg `overwrite_x` is only supported with its default value.
-
     """
     _check_plan(plan)
-    _check_overwrite_x(overwrite_x)
     x = _validate_input(x)
     norm = _swap_direction(norm)
     fsc = _compute_fwd_scale(norm, n, x.shape[axis])
 
     with _Workers(workers):
+        # Note: overwrite_x is not utilized
         result = mkl_fft.rfft(x, n=n, axis=axis, fwd_scale=fsc)
 
     np.conjugate(result, out=result)
@@ -569,10 +558,6 @@ def hfft2(
     Compute the 2-D FFT of a Hermitian complex array.
 
     For full documentation refer to `scipy.fft.hfft2`.
-
-    Limitation
-    -----------
-    The kwarg `overwrite_x` is only supported with its default value.
 
     """
     return hfftn(
@@ -600,10 +585,6 @@ def ihfft2(
     Compute the 2-D inverse FFT of a real spectrum.
 
     For full documentation refer to `scipy.fft.ihfft2`.
-
-    Limitation
-    -----------
-    The kwarg `overwrite_x` is only supported with its default value.
 
     """
     return ihfftn(
@@ -633,21 +614,17 @@ def hfftn(
 
     For full documentation refer to `scipy.fft.hfftn`.
 
-    Limitation
-    -----------
-    The kwarg `overwrite_x` is only supported with its default value.
-
     """
     _check_plan(plan)
-    _check_overwrite_x(overwrite_x)
     x = _validate_input(x)
     norm = _swap_direction(norm)
     x = np.array(x, copy=True)
     np.conjugate(x, out=x)
-    s, axes = _cook_nd_args(x, s, axes, invreal=True)
+    s, axes = _init_nd_shape_and_axes(x, s, axes, invreal=True)
     fsc = _compute_fwd_scale(norm, s, x.shape)
 
     with _Workers(workers):
+        # Note: overwrite_x is not utilized
         return mkl_fft.irfftn(x, s, axes, fwd_scale=fsc)
 
 
@@ -666,19 +643,15 @@ def ihfftn(
 
     For full documentation refer to `scipy.fft.ihfftn`.
 
-    Limitation
-    -----------
-    The kwarg `overwrite_x` is only supported with its default value.
-
     """
     _check_plan(plan)
-    _check_overwrite_x(overwrite_x)
     x = _validate_input(x)
     norm = _swap_direction(norm)
-    s, axes = _cook_nd_args(x, s, axes)
+    s, axes = _init_nd_shape_and_axes(x, s, axes)
     fsc = _compute_fwd_scale(norm, s, x.shape)
 
     with _Workers(workers):
+        # Note: overwrite_x is not utilized
         result = mkl_fft.rfftn(x, s, axes, fwd_scale=fsc)
 
     np.conjugate(result, out=result)
@@ -696,21 +669,19 @@ def get_workers():
 
 
 @contextlib.contextmanager
-def set_workers(n_workers):
+def set_workers(workers):
     """
     Set the value of workers used by default, returns the previous value.
 
     For full documentation refer to `scipy.fft.set_workers`.
 
     """
-    nw = operator.index(n_workers)
+    nw = operator.index(workers)
     token = None
     try:
         new_wd = _workers_data(nw)
         token = _workers_global_settings.set(new_wd)
         yield
     finally:
-        if token:
+        if token is not None:
             _workers_global_settings.reset(token)
-        else:
-            raise ValueError